Sprint 2: Basic Combat Elements
The games main combat mechanic is shooting/aiming. 
The dash from sprint 3 was implemented in this sprint to help test the "feel" of the game.
No enemies have been implemented.
The "shield" element will probably be implemented later.
The testing area's size increased.
After this point, commits won't be sprint-by-sprint since the code being written will be unfamiliar to me.